Update generatorRR5.generateUseRedirectFile to handle named exports
